chore: 完善日记管理功能以及一些配置
1.日记需要进行审核才能展示（审核后远程调用用户服务发送邮件通知
2.提取邮箱枚举类到公共代码中
3.新增拒绝和通过日记的邮箱策略

+++ text-service/src/main/java/com/llpy/textservice/entity/Diary.java
@@ -39,7 +39,7 @@ public class Diary {
     /**
      * 审核人
      */
-    private int passUser;
+    private Long passUser;
 
     private LocalDateTime createTime;
